+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 v0.2.3:
   * Clear the read buffer when sending an instruction.
+  * Add trace log for discarded partial messages after a timeout.
 
 v0.2.2:
   * Add debug and trace logs for skipped noise bytes.